Make IRenderer an abstract class
