+++ src/c-sharp/Infra/Factories/RendererFactory.cs
@@ -1,4 +1,4 @@
-using Archlens.Domain.Interfaces;
+using Archlens.Domain;
 using Archlens.Domain.Models.Enums;
 using Archlens.Domain.Models.Records;
 using Archlens.Infra.Renderers;
@@ -8,7 +8,7 @@ namespace Archlens.Infra.Factories;
 
 public sealed class RendererFactory
 {
-    public static IRenderer SelectRenderer(RenderOptions options) => options.Format switch
+    public static Renderer SelectRenderer(RenderOptions options) => options.Format switch
     {
         RenderFormat.Json => new JsonRenderer(),
         RenderFormat.PlantUML => new PlantUMLRenderer(),
